Added a config module that keeps track of the configurations
config.py: this module contains Config class. Config class has the following groups of attributes: 1) Paths related to the experiment, 2) Experiment configurations like experiment id. 3) Environment specs.
Exvironment specs are mostly related to the setup of SymPy.

Upon instantiation, Config class checks if there exists a Dumped directory or not. If not, it will be created. Then, Config class generates an experiment id with get_exp_id method. A directory with the name of experiment id will be created in the  Dumped directory.