Added the utils module that is responsible for building and performing environmental functions
utils.py: this module contains the class EnvHandler. This class builds an environment for the current experiment conatining attributes like vocabular, word2id and id2word. It has methods that are necessary for training and evaluations. One example is prefix_to_infox which takes a prefix representation of a mathematical expresession and outputs its infix representation. The infix representation then goes through infix_to_sympy to produce a sympy expression.

Most of the methods and attributes of this class have been taken from the authors original implementation.