Added some experiment-related and training-related features to the Config class
Experiment-related features: reload_size, which determines how many samples we are going to load from the training dataset. env_seed, which set the seed for the current experiment.

Traininf_related features: epoch_size: determines the size of each epooch during the training. Note: in the original implementation, authors did not set the epooch size to be the training data size. Instead, fro each epooch they get their samples from the dataset with a rendom index. batch_size: a feature that defines the batch size. learning_rate: a feature that stores the value of learning_rate