Added a module named datahandler which is responsible for loading and giving the data
This module has a class named MathData. MathData takes the config, environment, and dtype, which can be either 'train', 'valid, or 'test', and loads the data from the disk. This class inherited form torch.utils.data.Dataset, and overwrited __len__ and __getitem__ methods. It provides a method called collate_fn that gets lines from the dataset and returns a batch.

This module also has a function named ceate_data_loader, which builds and returns an iterable around the dataset, that we can get batches from.