Wraped the main process into the main function
Deleted CUDA data at the end of trainer and evaluator to free the cuda memory.
Added an input parser to main.py to parse the main configurations from the command line.
Altered Config class to accept a parser object and initialize its parameters with it.
Added a function to utils to parse the bolian variables.
Completed build_model function in Model/__init__.py