Wraped the main process into the main function
Added a command line parser to main.py to get the main configurations from user.
Altered Config class to accept a parser object as its input and initialize its parameters with it.
Deleted every unnecessary data in CUDA memory in trainer and evaluater to make the program compatible with low-memory GPU types.
Added a bool_flag to utils to handle input binary variables.
Fixed the build_model in Model/__init__ to create the model, load it (if necessary), and transfer it to config.device